# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookmarkHive API",
        "description": "Decentralized social bookmarking service based on the Activity Pub protocol.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bookmarkhive.test",
            "description": "API Dev Server"
        },
        {
            "url": "https://api2.bookmarkhive.test",
            "description": "API 2 Dev Server"
        }
    ],
    "paths": {
...
        "/users/me/following": {
            "get": {
                "tags": [
                    "Following"
                ],
                "summary": "List accounts being followed",
                "description": "Returns a paginated collection of accounts that the authenticated user is following. Default page size is 100 items, ordered by creation date (newest first).",
                "operationId": "listOwnFollowing",
                "parameters": [
                    {
                        "name": "after",
                        "in": "query",
                        "description": "Cursor for pagination - account ID to fetch results after",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of followed accounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collection": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FollowingShowPublic"
                                            }
                                        },
                                        "prevPage": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "nextPage": {
                                            "description": "URL for next page if available",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "following_list": {
                                        "summary": "Paginated list of followed accounts",
                                        "value": {
                                            "collection": [
                                                {
                                                    "id": "01234567-89ab-cdef-0123-456789abcdef",
                                                    "account": {
                                                        "@iri": "https://bookmarkhive.test/profile/janedoe",
                                                        "username": "janedoe",
                                                        "instance": "bookmarkhive.test",
                                                        "inboxUrl": "https://bookmarkhive.test/profile/janedoe/inbox",
                                                        "outboxUrl": "https://bookmarkhive.test/profile/janedoe/outbox",
                                                        "sharedInboxUrl": "https://bookmarkhive.test/inbox",
                                                        "followerUrl": "https://bookmarkhive.test/profile/janedoe/followers",
                                                        "followingUrl": "https://bookmarkhive.test/profile/janedoe/following"
                                                    },
                                                    "createdAt": "2024-01-01T12:00:00+00:00"
                                                }
                                            ],
                                            "prevPage": null,
                                            "nextPage": "https://bookmarkhive.test/users/me/following?after=01234567-89ab-cdef-0123-456789abcdef",
                                            "total": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/following/{usernameWithInstance}": {
            "post": {
                "tags": [
                    "Following"
                ],
                "summary": "Follow an account",
                "description": "Creates a follow relationship with the specified account by username.",
                "operationId": "followAccount",
                "parameters": [
                    {
                        "name": "usernameWithInstance",
                        "in": "path",
                        "description": "Username of the account to follow (with instance)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "janedoe@bookmarkhive.test"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account followed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FollowingShowPublic"
                                },
                                "examples": {
                                    "followed_account": {
                                        "summary": "Successfully followed account",
                                        "value": {
                                            "id": "01234567-89ab-cdef-0123-456789abcdef",
                                            "account": {
                                                "@iri": "https://bookmarkhive.test/profile/janedoe",
                                                "username": "janedoe",
                                                "instance": "bookmarkhive.test",
                                                "inboxUrl": "https://bookmarkhive.test/profile/janedoe/inbox",
                                                "outboxUrl": "https://bookmarkhive.test/profile/janedoe/outbox",
                                                "sharedInboxUrl": "https://bookmarkhive.test/inbox",
                                                "followerUrl": "https://bookmarkhive.test/profile/janedoe/followers",
                                                "followingUrl": "https://bookmarkhive.test/profile/janedoe/following"
                                            },
                                            "createdAt": "2024-01-01T12:00:00+00:00"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "409": {
                        "description": "Conflict - follow request already exist"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Following"
                ],
                "summary": "Unfollow an account",
                "description": "Removes the follow relationship with the specified account by username (with instance).",
                "operationId": "unfollowAccount",
                "parameters": [
                    {
                        "name": "usernameWithInstance",
                        "in": "path",
                        "description": "Username of the account to unfollow",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "janedoe@bookmarkhive.test"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Account unfollowed successfully"
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AccountShowPublic": {
                "description": "Account object representing an ActivityPub account",
                "properties": {
                    "@iri": {
                        "description": "IRI of the account resource",
                        "type": "string",
                        "format": "iri"
                    },
                    "username": {
                        "description": "Account username",
                        "type": "string"
                    },
                    "instance": {
                        "description": "Instance domain where the account is hosted",
                        "type": "string"
                    },
                    "inboxUrl": {
                        "description": "ActivityPub inbox URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "outboxUrl": {
                        "description": "ActivityPub outbox URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "sharedInboxUrl": {
                        "description": "ActivityPub shared inbox URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "followerUrl": {
                        "description": "ActivityPub followers collection URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "followingUrl": {
                        "description": "ActivityPub following collection URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "FollowingShowPublic": {
                "description": "Following object representing an account that the user follows",
                "properties": {
                    "id": {
                        "description": "Following relationship ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "account": {
                        "$ref": "#/components/schemas/AccountShowPublic"
                    },
                    "createdAt": {
                        "description": "When the follow relationship was created",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from POST /auth endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}